backend/requirements.txt
# ============================================================================
# TruthLens AI - Production Backend Requirements
# ============================================================================
# This file contains all dependencies for the TruthLens AI backend.
# Use: pip install -r requirements.txt
# 
# For a minimal installation without heavy dependencies, use:
# pip install -r requirements-simple.txt
# ============================================================================

# === Core Framework & API ===
fastapi==0.104.1
uvicorn==0.24.0
starlette==0.27.0
python-multipart==0.0.6

# === Web Server & Async ===
aiofiles==23.2.1
python-dotenv==1.0.0
requests==2.31.0

# === Database & ORM ===
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# === API Validation & Data ===
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic-extra-types==2.1.0

# === Security & Authentication ===
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
cryptography==41.0.7
PyJWT==2.12.1

# === ML/NLP Core Models ===
torch==2.1.1
transformers==4.57.6
huggingface-hub==0.36.2
sentence-transformers==5.1.2
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2

# === Data Processing ===
numpy==1.26.2
pandas==2.1.3
scipy==1.11.4

# === Network & Graph Analysis ===
networkx==3.2.1

# === Explainability & Interpretability ===
shap==0.43.0
lime==0.2.0

# === Image Processing (Optional) ===
opencv-python==4.8.1.78
Pillow==10.1.0
pytesseract==0.3.10
scikit-image==0.22.0

# === Web Scraping & Content Extraction ===
newspaper3k==0.2.8
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10

# === Vector Databases ===
pinecone-client==3.0.1  # DEPRECATED - replaced with new 'pinecone' package below
pinecone==3.0.0  # Latest Pinecone SDK (replaces pinecone-client)
faiss-cpu==1.7.4

# === Caching & Async Processing ===
redis==5.0.1
celery==5.3.4
aioredis==2.0.1

# === Cloud Storage ===
boto3==1.29.7

# === Monitoring & Logging ===
prometheus-client==0.19.0
sentry-sdk==1.38.0

# === Testing ===
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

# === Utilities ===
pytz==2023.3.post1
requests-cache==1.1.0
tenacity==8.2.3
tqdm==4.66.1

